Only add `--no-run-if-empty` if GNU xargs is detected.
This means that findx now works with the native BSD xargs, which
implements the `--no-run-if-empty` behavior by default--despite POSIX
saying the tool should be run at least once.

I went the detection route because it is possible that GNU grep is
present, if someone installed grep with Homebrew using the
`--default-names` option.